# Reduce memory overhead for webpack configs with many entries