<!--
  HOLY CRAP a Pull Request. We ❤️ those!

  If you remove or skip this template, you'll make the 🐼 sad and the mighty god
  of Github will appear and pile-drive the close button from a great height
  while making animal noises.

  Please place an x (no spaces!) in all [ ] that apply
-->

This PR contains a:

- [x] **bugfix**
- [ ] new **feature**
- [x] **code refactor**
- [ ] **test update** <!-- if bug or feature is checked, this should be too -->
- [ ] **typo fix**
- [ ] **metadata update**

### Motivation / Use-Case

This change dramatically reduces the total size of objects on the heap at any one time for webpack configurations with many entries. Because `tasks` used to be computed eagerly and `completedTasks` was collected as a large Array, space complexity used to be linear with respect to the number of entries. (More precisely, the max heap size was proportional to the combined size in bytes of all pre-minified sources).

Now, we defer the generation of a `task` (and thus the memory allocation required for computing `asset.source()` of an entry) until a worker is made available. Similarly, the computation of `serialize(task)`, another large String, is deferred until a worker is available. Finally, when a
`task` is completed, the `completedTask` is consumed immediately, releasing the reference to the original `asset.source()`, and making it a candidate for garbage collection.

The effect is that space complexity is now roughly linear with respect to the number of parallel workers. The number of pre-minified sources that are considered "live" to the garbage collector will never exceed the number of parallel workers.

I believe this fixes #143 for most people experiencing the problem.

I created a repository that reliably reproduces the issue: https://github.com/cjlarose/terser-webpack-plugin-out-of-memory. Using the changes introduced in this PR, that project builds successfully using the default Node old space size (1400MB).

### Breaking Changes

None

### Additional Info