@cjlarose 2020-01-15
I collected benchmarks using the repo I created that demonstrates the original `JavaScript heap out of memory` bug (https://github.com/cjlarose/terser-webpack-plugin-out-of-memory/tree/cf9fc24cee6b2085a472de3b36a37be7210238e7)

I executed builds using my Mid-2017 MacBook Pro (Intel Core i7-7820HQ Quad-Core 2.9GHz), with cache off and cache on, with 4 terser workers. I'm using macOS 10.14.6 and Node 12.14.1.

"Fails" in this table represents somewhere where `terser-webpack-plugin` version 2.3.2 failed with an out-of-memory error.


| # entries | cache | max-old-space-size | 2.3.2 (s) | this PR  (s) |
|-----------|-------|--------------------|-----------|--------------|
| 10        | cold  | 1400MB             | 4.17      | 4.76         |
| 10        | warm  | 1400MB             | 1.05      | 1.19         |
| 100       | cold  | 1400MB             | 23.15     | 23.76        |
| 100       | warm  | 1400MB             | 1.77      | 1.98         |
| 500       | cold  | 1400MB             | Fails     | 104.42       |
| 500       | warm  | 1400MB             | 5.10      | 6.38         |
| 500       | cold  | 4096MB             | 109.32    | 103.79       |
| 500       | warm  | 4096MB             | 5.13      | 6.02         |
| 1000      | cold  | 1400MB             | Fails     | 214.10       |
| 1000      | warm  | 1400MB             | 10.50     | 10.42        |
| 1000      | cold  | 4096MB             | Fails     | 242.01       |
| 1000      | warm  | 4096MB             | 10.68     | 11.53        |

In conclusion, there are some configurations for which this change introduces a minor performance regression. This is mostly expected since we're basically forcing the "main" process to wait until work is completed before generating the next task in order to conserve memory. The takeaway, though, is that this change makes it possible to build webpack configurations with large numbers of entries reliably.