@avshalomt2 2020-01-21
@cjlarose Do you believe this fix solves the same error when using `terser-webpack-plugin-legacy` (based on branch `webpack-3`)?
I just looked at the code in the `webpack-3` branch and it seems possible to backport this fix to there, and i'm thinking about doing so.

@evilebottnawi If I create such a PR, could it be merge to the currently stale `webpack-3` branch?