@evilebottnawi 2020-01-22
> If I create such a PR, could it be merge to the currently stale webpack-3 branch?

No, please migrate on latest version