@evilebottnawi 2020-01-28
I rewrited your idea like JavaScript Beginner Book :smile: Big thanks for a inspiration! It was great idea. Perhaps you still have some ideas for optimization, so I’ll leave this PR open for a while, feel free to leave feedback