@cjlarose 2020-01-28
I'll close it for now. I can introduce separate optimizations in new PRs if necessary. Thanks.