@cjlarose 2020-01-30
This work was re-implemented in #211